% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Compute the Means squared error}
\usage{
mse(truth, predicted)
}
\arguments{
\item{truth}{A vector of numbers representing the observed values}

\item{predicted}{A vector of predicted outcomes}
}
\value{
An integer representing the Mean squared error
}
\description{
This package allows you to direclty compute the mean squared error,
making it easier to evaluate your model performance.
}
\examples{
truth<-c(1,2,3,4,5)
predicted<-c(0,0,2,3,4)
mse(truth,predicted)

}
\author{
Marta Sech \email{m.sech@students.uu.nl}
}
